refactor(tests): Moved email local tests
