+++ test/local/email/delivery_tests.js
@@ -6,8 +6,8 @@ const assert = require('insist')
 
 const EventEmitter = require('events').EventEmitter
 const sinon = require('sinon')
-const spyLog = require('../mocks').spyLog
-const delivery = require('../../lib/email/delivery')
+const spyLog = require('../../mocks').spyLog
+const delivery = require('../../../lib/email/delivery')
 
 const mockDeliveryQueue = new EventEmitter()
 mockDeliveryQueue.start = function start() {
